remove prometheus pushgateway
