+++ flink-metrics/flink-metrics-prometheus/src/test/java/org/apache/flink/metrics/prometheus/PrometheusReporterTest.java
@@ -178,7 +178,7 @@ public class PrometheusReporterTest extends TestLogger {
 
 	@Test
 	public void doubleGaugeIsConvertedCorrectly() {
-		assertThat(reporter.gaugeFrom(new Gauge<Double>() {
+		assertThat(PrometheusReporter.gaugeFrom(new Gauge<Double>() {
 			@Override
 			public Double getValue() {
 				return 3.14;
@@ -188,7 +188,7 @@ public class PrometheusReporterTest extends TestLogger {
 
 	@Test
 	public void shortGaugeIsConvertedCorrectly() {
-		assertThat(reporter.gaugeFrom(new Gauge<Short>() {
+		assertThat(PrometheusReporter.gaugeFrom(new Gauge<Short>() {
 			@Override
 			public Short getValue() {
 				return 13;
@@ -198,7 +198,7 @@ public class PrometheusReporterTest extends TestLogger {
 
 	@Test
 	public void booleanGaugeIsConvertedCorrectly() {
-		assertThat(reporter.gaugeFrom(new Gauge<Boolean>() {
+		assertThat(PrometheusReporter.gaugeFrom(new Gauge<Boolean>() {
 			@Override
 			public Boolean getValue() {
 				return true;
@@ -211,7 +211,7 @@ public class PrometheusReporterTest extends TestLogger {
 	 */
 	@Test
 	public void stringGaugeCannotBeConverted() {
-		assertThat(reporter.gaugeFrom(new Gauge<String>() {
+		assertThat(PrometheusReporter.gaugeFrom(new Gauge<String>() {
 			@Override
 			public String getValue() {
 				return "I am not a number";
